move file from isolated
