+++ packages/flutter_tools/lib/src/isolated/devfs_web.dart
@@ -29,8 +29,8 @@ import '../vmservice.dart';
 import '../web/bootstrap.dart';
 import '../web/chrome.dart';
 import '../web/compile.dart';
+import '../web/devfs_config.dart';
 import '../web_template.dart';
-import 'devfs_config.dart';
 import 'web_asset_server.dart';
 
 const kLuciEnvName = 'LUCI_CONTEXT';
