+++ packages/flutter_tools/lib/src/isolated/resident_web_runner.dart
@@ -37,12 +37,11 @@ import '../run_hot.dart';
 import '../vmservice.dart';
 import '../web/chrome.dart';
 import '../web/compile.dart';
+import '../web/devfs_config.dart';
 import '../web/file_generators/flutter_service_worker_js.dart';
 import '../web/file_generators/main_dart.dart' as main_dart;
 import '../web/web_device.dart';
 import '../web/web_runner.dart';
-
-import 'devfs_config.dart';
 import 'devfs_web.dart';
 import 'web_expression_compiler.dart';
