+++ packages/flutter_tools/lib/src/isolated/web_asset_server.dart
@@ -28,11 +28,11 @@ import '../globals.dart' as globals;
 import '../web/bootstrap.dart';
 import '../web/chrome.dart';
 import '../web/compile.dart';
+import '../web/devfs_config.dart';
+import '../web/devfs_proxy.dart';
 import '../web/memory_fs.dart';
 import '../web/module_metadata.dart';
 import '../web_template.dart';
-import 'devfs_config.dart';
-import 'devfs_proxy.dart';
 import 'release_asset_server.dart';
 import 'web_server_utlities.dart';
