+++ packages/flutter_tools/test/general.shard/web/proxy_test.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 
 import 'package:flutter_tools/src/globals.dart' as globals;
-import 'package:flutter_tools/src/isolated/devfs_proxy.dart';
+import 'package:flutter_tools/src/web/devfs_proxy.dart';
 import 'package:shelf/shelf.dart';
 import 'package:test/test.dart';
 import 'package:yaml/yaml.dart';
